fastapi-redis-postgresql-kubernetes/v1.0/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.25
async-exit-stack==1.0.1
psycopg2==2.9.1
pydantic==1.9.0
psycopg2-binary==2.9.3
aioredis==2.0.0
fastapi-cache2
redis



